requirements.txt
flask==2.3.3
requests
python-binance
python-telegram-bot==13.15
python-dotenv